updated Finance page
