+++ src/SFA.DAS.SupportConsole.UITests/Project/Tests/Features/MA/Tier2/SC_T2_01.feature
@@ -5,7 +5,7 @@
 @supportconsole
 @masupportconsole
 Scenario: SC_T2_01_Tier2Navigation
-	Given the Tier 1 User is logged into Support Console
+	Given the Tier 2 User is logged into Support Console
 	And the User is on the Account details page
 	When the user navigates to finance page
 	Then the user is redirected to finance page
