+++ src/SFA.DAS.SupportConsole.UITests/Project/Tests/Pages/FinancePage.cs
@@ -1,12 +1,11 @@
-﻿using DnsClient.Protocol;
-using OpenQA.Selenium;
+﻿using OpenQA.Selenium;
 using TechTalk.SpecFlow;
 
 namespace SFA.DAS.SupportConsole.UITests.Project.Tests.Pages
 {
     public class FinancePage : SupportConsoleBasePage
     {
-        protected override string PageTitle => "Finanace";
+        protected override string PageTitle => "Finance";
 
         protected override By PageHeader => By.CssSelector(".heading-xlarge");
 
@@ -20,13 +19,9 @@ namespace SFA.DAS.SupportConsole.UITests.Project.Tests.Pages
 
         private By CurrentBalance => By.CssSelector(".data__purple-block");
 
-        public FinancePage(ScenarioContext context, bool navigate = false) : base(context)
+        public FinancePage(ScenarioContext context) : base(context)
         {
             _context = context;
-            if (navigate)
-            {
-                ClickFinanceMenuLink();
-            }
             VerifyPage();
         }
 
@@ -42,31 +37,5 @@ namespace SFA.DAS.SupportConsole.UITests.Project.Tests.Pages
             _pageInteractionHelper.VerifyText(CurrentBalance, "Current balance");
         }
 
-    }
-
-    public class LevyDeclarationsPage : SupportConsoleBasePage
-    {
-
-        protected override string PageTitle => "Levy declarations";
-
-        protected override By PageHeader => By.CssSelector(".heading-large");
-
-        public LevyDeclarationsPage(ScenarioContext context) : base(context)
-        {
-            VerifyPage();
-        }
-    }
-
-    public class TeamMembersPage : SupportConsoleBasePage
-    {
-
-        protected override string PageTitle => "Team members";
-
-        protected override By PageHeader => By.CssSelector(".heading-xlarge");
-
-        public TeamMembersPage(ScenarioContext context) : base(context)
-        {
-            VerifyPage();
-        }
     }
 }
